Create leoAPI.py. Add tests for icon bar API
